refactor(tests): simplify mock setup and error handling
